# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: InvestKE Investment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions/deposit:
    post:
      operationId: initiateDeposit
      tags: [transactions]
      summary: Initiate M-Pesa deposit (STK push)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "200":
          description: STK push sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositResponse"
  /transactions/withdraw:
    post:
      operationId: requestWithdrawal
      tags: [transactions]
      summary: Request a withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Withdrawal requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
  /admin/transactions:
    get:
      operationId: adminGetTransactions
      tags: [admin]
      summary: Admin - list all transactions
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTransactions"
...
components:
  schemas:
...
    Transaction:
      type: object
      required: [id, userId, type, amount, fee, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [deposit, withdrawal, earning]
        amount:
          type: number
        fee:
          type: number
          description: Processing fee deducted (withdrawals only)
        status:
          type: string
          enum: [pending, completed, failed, rejected]
        method:
          type: ["string", "null"]
          enum: [mpesa, airtel_money, bank, null]
        phoneOrAccount:
          type: ["string", "null"]
        reference:
          type: ["string", "null"]
        isFlagged:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    DepositInput:
      type: object
      required: [amount, phone]
      properties:
        amount:
          type: number
          description: Amount in KES
        phone:
          type: string
          description: M-Pesa phone number
    DepositResponse:
      type: object
      required: [message, transactionId]
      properties:
        message:
          type: string
        transactionId:
          type: integer
        checkoutRequestId:
          type: ["string", "null"]
    WithdrawalInput:
      type: object
      required: [amount, method, phoneOrAccount]
      properties:
        amount:
          type: number
        method:
          type: string
          enum: [mpesa, airtel_money, bank]
        phoneOrAccount:
          type: string
          description: Phone number for mobile money or account number for bank
...
    PaginatedTransactions:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...